Refactorizare AuthController
