+++ Final Project/Server/fiimentor/controllers/AuthController/index.ts
@@ -1,0 +1,5 @@
+import {login} from './Login'
+import {registerRole} from './RegisterRole';
+import {register} from './Register';
+
+export {login,registerRole,register};
